+++ compiler/src/org.graalvm.compiler.truffle.runtime/src/org/graalvm/compiler/truffle/runtime/GraalRuntimeSupport.java
@@ -304,8 +304,8 @@ final class GraalRuntimeSupport extends RuntimeSupport {
     }
 
     @Override
-    public Object[] getNonPrimitiveResolvedFields(Class<?> type) {
-        return GraalTruffleRuntime.getRuntime().getNonPrimitiveResolvedFields(type);
+    public Object[] getResolvedFields(Class<?> type, boolean includePrimitive, boolean includeSuperclasses) {
+        return GraalTruffleRuntime.getRuntime().getResolvedFields(type, includePrimitive, includeSuperclasses);
     }
 
     @Override